test/e2e/requirements.txt
aiofiles==23.2.1
altair==5.1.2
annotated-types==0.6.0
anyio==4.0.0
apache-skywalking==1.0.1
astor==0.8.1
astroid==2.11.7
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
darglint==1.8.1
dill==0.3.7
eradicate==2.3.0
fastapi==0.103.0
ffmpy==0.3.1
filelock==3.13.1
flake8==5.0.4
flake8-2020==1.8.1
flake8-bugbear==22.12.6
flake8-comprehensions==3.14.0
flake8-docstrings==1.7.0
flake8-eradicate==1.5.0
flake8-use-fstring==1.4
Flask==2.3.3
flynt==0.76
fonttools==4.44.0
fsspec==2023.10.0
gradio==3.50.2
gradio_client==0.6.1
grpcio==1.65.4
grpcio-tools==1.59.2
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
huggingface-hub==0.18.0
idna==3.4
importlib-resources==6.1.1
inflect==6.2.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpickle==3.0.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
matplotlib==3.7.3
mccabe==0.7.0
nltk==3.8.1
numpy==1.24.4
orjson==3.9.10
packaging==23.2
pandas==2.0.3
pep8-naming==0.13.3
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.3.0
protobuf==4.25.0
psutil==5.9.6
pycodestyle==2.9.1
pydantic==2.4.2
pydantic_core==2.10.1
pydocstyle==6.3.0
pydub==0.25.1
pyflakes==2.5.0
pylint==2.13.9
pyparsing==3.1.1
pytest==7.4.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
-e git+https://github.com/mrproliu/R3.git@4ae163afc991596c26bc774a0da143baf9e0d376#egg=R3
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rpds-py==0.12.0
semantic-version==2.10.0
setuptools-scm==8.0.4
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
starlette==0.27.0
tomli==2.0.1
toolz==0.12.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
unify==0.5
untokenize==0.1.1
urllib3==2.0.7
uvicorn==0.24.0.post1
websockets==11.0.3
Werkzeug==3.0.1
wrapt==1.15.0
